+++ client/src/test/scala/cakesolutions/kafka/KafkaSpec.scala
@@ -1,4 +1,4 @@
-package cakesolutions
+package cakesolutions.kafka
 
 import org.scalatest.{FlatSpec, Matchers}
